+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -27,6 +27,7 @@ CHANGELOG
  * deprecated commands auto registration
  * added `AddCacheClearerPass`
  * added `AddCacheWarmerPass`
+ * deprecated `Bundle::registerCommands()`
 
 3.3.0
 -----